requirements.txt
python-dotenv >= 0.19, < 0.20mattermostdriver >= 7.3.2, < 7.4.0